Switch to elm-route-url
This kinda works but:
- hashes look ugly and I can't parse them with my existing parser
-maybe I can prse them using RouteUrl.Builders hash functions
but they're still going to look ugly

- And if I use propper paths I can't get it to work on url changes
if the path includes a slash, which it is interpreting as a different
resouce location even with --pushstate in the command line.
So maybe we need a different way of servin up the html page
using a real web server or something?