Calculate neighbours in update not view
Cache the prev and next neighbours in the page model
and fill them in when we visit a page -- for a given
set of entries. That means when we get more entries, we also
have to recalculate the neighboubrs in case they have
changed as a result. Which turns out to be the case for
the initital load when we try to load the page
but we don't yet have any entris so we dont ungrey the
arrow buttons.