Fix bug fetching non adjacent later items
When fetching later items I was not odering the results ascending so I
wasnt being sent the adjacent ones and I ended up with gaps in the
listing.

To add the ascending parameter I need to add another query parameter
to the http request for which I used HttpBuilder to biuld the requests.