Fix recuring problem with duplicate entries
So we were still gettting duplicate entries but this time in GMT
when fromIsoString was assuming BST for summer time dates
even when we knew they were really GMT (UTC) which is kinda
correct because ISO8601 should be interpreted as local dates unless
you add a Z. So now I'm forcing those GMT date strings to end
wtih Z before I parse them.  This appears to have removed the
duplicate entries but I notice that the displayed  times
are now out, so it looks like the formatted output is
appluing a timezone from somewhere.